Move database to its own file
